Fix font size inconsistencies
Since many (basically all) things are based on `em` units, there are
strange interactions between the font size and other components of the
template. For instance, since the margins of the slides were `em`-based,
they were affected by the text size in the slide itself.

In order to avoid such problems, font size is normalized (there is a
default font size and you can specify a new one, globally or per slide)
and things such as margins (which should a priori remain fixed) are
defined in absolute units instead.